fix: remove game_ctx from pfr_play, fix savestate, fix 64-bit crashes
The game_ctx system (multi-env context struct for RL training) was half-
integrated into pfr_play.c, making standalone builds impossible. This
commit removes game_ctx from pfr_play while keeping it available for
libpfr_game.so.

What broke and why:
- pfr_play.c had 65 `g_ctx->` dereferences replaced with direct global
  access. game_ctx_alloc/free calls removed from boot sequence.
- game_ctx.h stripped to includes-only + dummy typedef. Macros header
  made a no-op. This lets gen files compile without the macro system.
- BuildSegmentTable in host_savestate.c used dl_iterate_phdr to find
  memory segments -- this works for .so but FAILS for the main
  executable (dladdr returns base=0, phdr search misses). Added
  fallback using linker symbols (__data_start/_edata/__bss_start/_end).
  Savestates were "succeeding" while capturing 0 bytes.
- inject_game_ctx.py was disabled in CMakeLists.txt but gen files still
  had stale game_ctx.h/#include injections from prior builds. Made both
  generated headers no-ops so stale includes are harmless.
- Asset pipeline was reading from gen/ (stale copies) instead of
  upstream sources for files with HOST_NATIVE guards, stripping the
  guards. Fixed 7 files in CMakeLists.txt to read from upstream.

64-bit pointer crashes fixed:
- FieldEffectScript_ReadPtr: script bytecode stores 32-bit indices into
  pointer lookup table, but callers cast raw u32 to pointers. Added
  ReadPtr variant using T2_READ_PTR -> HostScriptPtrLookup.
- ScrCmd_applymovement: used ScriptReadWord (raw u32 index 0x2b39)
  instead of ScriptReadPtr. SIGSEGV on Rock Smash / any movement script.
- region_map.c: sMapCursor NULL deref (fault_addr=0x2) when opening
  Fly map. Cursor allocated lazily; on GBA addr 0x2 is valid BIOS RAM.
- field_effect.c: vblankCallback truncated to u32 via
  StoreWordInTwoHalfwords. Replaced with HostPtrStore_Put/Get.
- new_menu_helpers.c: (u32)ptr truncated heap pointer. Changed to
  (uintptr_t)ptr.

Features added:
- F3 egg hatch auto-walk with info panel overlay
- --load-state CLI option for direct .pfrstate boot
- F1/F2 savestate status messages in terminal
- HostRendererRenderFrame after state restore for immediate visual update

+++ src/host_intro_stubs.c
@@ -32,7 +32,7 @@ u32 gHostIntroStubResetBgPositionsCalls = 0;
 u32 gHostIntroStubStartBlendTaskCalls = 0;
 u32 gHostIntroStubIsBlendTaskActiveCalls = 0;
 /* m4aSongNumStart, PlaySE, PlayCry_ByMode now upstream */
-u8 gHostIntroStubLoadGameSaveResult = SAVE_STATUS_EMPTY;
+u32 gHostIntroStubLoadGameSaveResult = SAVE_STATUS_EMPTY;
 /* gSaveFileStatus now from save.c */
 
 __asm__(
